requirements.txt
Django
pillow
django-bootstrap5
mysqlclient
django-filter